[Feat] Setup Reward depending of room Element

+++ Assets/Scripts/A Trier/TerrainGenerator.cs
@@ -287,7 +287,7 @@ public class TerrainGenerator : MonoBehaviour
             TeleporterFeebackController tpFeedback = currentRoomManager.teleporterArray[i].GetComponentInChildren<TeleporterFeebackController>();
             tpFeedback.rewardToUse = (int)roomManager.rewardType;
             tpFeedback.ChangeRewardID(tpFeedback.rewardToUse, roomManager.m_materialPreviewTRT);
-            tpFeedback.ChangeColorVFX(roomManager.element);
+            tpFeedback.ChangeColorVFX((int)roomManager.element);
         }
 
         currentRoomManager.SetupTeleporter(terrainInstantiated.Count);
